Attempt to debug Store and Load
There seems to be an issue with these memory operations and how they interact with the stack.
We found Load would pop a negative value from the stack and add an offset, ultimately getting
a very large number (roughly 2^32), which will then get a memory out of bounds error and fail
to push a value back onto the stack.